Extract SplashManager from WebManager
