+++ common/src/main/java/com/wynntils/gui/screens/WynntilsMenuScreen.java
@@ -5,7 +5,6 @@
 package com.wynntils.gui.screens;
 
 import com.mojang.blaze3d.vertex.PoseStack;
-import com.wynntils.core.webapi.WebManager;
 import com.wynntils.features.user.map.MapFeature;
 import com.wynntils.gui.render.FontRenderer;
 import com.wynntils.gui.render.HorizontalAlignment;
@@ -20,6 +19,7 @@ import com.wynntils.mc.objects.CustomColor;
 import com.wynntils.mc.utils.ComponentUtils;
 import com.wynntils.mc.utils.McUtils;
 import com.wynntils.wynn.model.CharacterManager;
+import com.wynntils.wynn.model.SplashManager;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
@@ -229,7 +229,7 @@ public class WynntilsMenuScreen extends WynntilsMenuScreenBase {
                         HorizontalAlignment.Center,
                         FontRenderer.TextShadow.NONE);
 
-        String currentSplash = WebManager.getCurrentSplash();
+        String currentSplash = SplashManager.getCurrentSplash();
         currentSplash = currentSplash == null ? "" : currentSplash;
         FontRenderer.getInstance()
                 .renderAlignedTextInBox(
